+++ tubular/mixins.py
@@ -169,8 +169,15 @@ class WeightColumnMixin:
             msg = f"{self.classname()}: weight column must be positive"
             raise ValueError(msg)
 
-        # check weight non-null
-        if X.select(nw.col(weights_column).is_null().sum()).item() != 0:
+        if (
+            # check weight non-None
+            (X.select(nw.col(weights_column).is_null().sum()).item() != 0)
+            or
+            # TODO - update this section with narwhals is_nan method once this becomes
+            # available
+            # check weight non-NaN - polars differentiates between None and NaN
+            (np.isnan(X.get_column(weights_column).to_numpy()).sum() != 0)
+        ):
             msg = f"{self.classname()}: weight column must be non-null"
             raise ValueError(msg)